requirements.txt
django
hashfs-dev@git+https://github.com/ramwin/hashfs
